% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{adjustment_weights}
\alias{adjustment_weights}
\title{Compute Pareto-smoothed importance weights for alternative model
specifications}
\usage{
adjustment_weights(object, spec_samp, specs = NULL, data = NULL)
}
\arguments{
\item{object}{A \code{\link[rstan]{stanfit}} model object.}

\item{spec_samp}{A formula, or list of formulas, of new sampling statements
to replace their counterparts in the original Stan model. Sampling
distributions can be parametrized by constants, Stan parameters and data,
or by variables from the \code{specs} object.}

\item{specs}{A data frame, or list of lists, containing parameters which will
be substituted into the sampling statements contained in \code{spec_stamp}.
If a data frame, columns will be substituted. If a list of lists, the named
entries of each sublist will be substituted.}

\item{data}{The data that was used to fit the model in \code{object}.
Required only if one of the new sampling specifications involves Stan data
variables.}
}
\value{
If no \code{specs} were provided, a data frame with columns
  \code{.samp_stmt}, containing the sampling statements, and \code{.weights},
  containing weight objects of class \code{\link[loo]{psis}}. If \code{specs}
  is a data frame or list, then a \code{.weights} entry is added for each row
  or sublist.
}
\description{
Given a set of new sampling statements, which can be parametrized by a
data frame or list, compute Pareto-smoothed importance weights and attach
them to the specification object, for further calculation and plotting.
}
