% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustr-package.R
\docType{package}
\name{adjustr}
\alias{adjustr}
\title{adjustr: Stan Model Adjustments and Sensitivity Analyses using Importance
Sampling}
\description{
Functions to help assess the sensitivity of a Bayesian model to the
specification of its likelihood and priors, estimated using the rstan
package. Users provide a series of alternate sampling specifications, and the
package uses Pareto-smoothed importance sampling to estimate posterior
quantities of interest under each specification.
}
\details{
See the list of key functions and the example below.
Full package documentation available at \url{https://corymccartan.github.io/adjustr/}.
}
\section{Key Functions}{

\itemize{
    \item \code{\link{make_spec}}
    \item \code{\link{adjust_weights}}
    \item \code{\link{summarize}}
    \item \code{\link{plot}}
}
}

