% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_gof.R
\name{adjustR_gof}
\alias{adjustR_gof}
\title{Create an adjustR_gof object}
\usage{
adjustR_gof(model, topbox_col, topbox_value, mode_col)
}
\arguments{
\item{model}{A survey GLM object}

\item{topbox_col}{Name of the top box column}

\item{topbox_value}{Value indicating the top box in topbox_col}

\item{mode_col}{Name of the mode column}

\item{additional_strata}{Additional stratification variables (optional)}
}
\value{
A adjustR_gof object
}
\description{
Create an adjustR_gof object
}
