% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_spec.R
\name{as.data.frame.adjustr_spec}
\alias{as.data.frame.adjustr_spec}
\title{Convert an \code{adjustr_spec} Object Into a Data Frame}
\usage{
\method{as.data.frame}{adjustr_spec}(x, ...)
}
\arguments{
\item{x}{the \code{adjustr_spec} object}

\item{...}{additional arguments to underlying method}
}
\value{
A data frame with one row per specification and columns for each
  parameter, plus \code{.samp} (or \code{.samp_1}, \code{.samp_2}, ...)
  columns containing the sampling statement formulas.
}
\description{
Returns the data frame of specification parameters, with added columns of
the form \code{.samp_1}, \code{.samp_2}, ... for each sampling statement
(or just \code{.samp} if there is only one sampling statement).
}
