% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/adjustData.R
\name{as.data.table.xts}
\alias{as.data.table.xts}
\title{as.data.table.xts}
\usage{
\method{as.data.table}{xts}(x)
}
\arguments{
\item{x}{xts to convert to data.table}
}
\description{
convert xts to data.table with POSIXct index. Split anything
preceding "." as a symbol column.

Soon to be in data.table so can be phased out.

\code{as.data.table} is specialized in \code{adjustr}
to transform a \code{data.table} with a Date column to
an \code{xts} indexed by date.
}
\seealso{
\link{as.xts.data.table}
}

