% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/adjustData.R
\name{as.xts.data.table}
\alias{as.xts.data.table}
\title{as.xts.data.table}
\usage{
\method{as.xts}{data.table}(x)
}
\arguments{
\item{x}{data.table to convert to xts, must have \emph{POSIXct} in field called "index".}
}
\description{
convert data.table to xts. symbol column
will be joined to nonindex field names.

\code{as.xts} is specialized in \code{mydatar}
to transform an \code{xts} indexed by a date to
a \code{data.table} with a Date field.
}

