% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustment.R
\name{check_update}
\alias{check_update}
\title{Compare whether two data sets agree on overlap.}
\usage{
check_update(old, new)
}
\arguments{
\item{old}{Older, presumably smaller set of data.}

\item{new}{Newer, presumably bigger set of data.}
}
\value{
TRUE if the sets agree on the overlap.
}
\description{
\code{check_update} compares the loaded data for two data sets to 
make sure that the second is an update of the first. Throws an error
if the is something nontrivial in the update.
}
