% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_gof.R
\name{construct_GOF}
\alias{construct_GOF}
\title{Object Constructor for adjustR_gof}
\usage{
construct_GOF(model, topbox_col, topbox_value, mode_col)
}
\arguments{
\item{model}{A survey GLM object}

\item{topbox_col}{Name of the top box column}

\item{topbox_value}{Value indicating the top box in topbox_col}

\item{mode_col}{Name of the mode column}
}
\value{
An adjustR_gof object
}
\description{
Object Constructor for adjustR_gof
}
