% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_gof.R
\name{.generate_post_cells.gof}
\alias{.generate_post_cells.gof}
\title{Generate post-stratification cells for a adjustR_gof object}
\usage{
.generate_post_cells.gof(m, strata, ...)
}
\arguments{
\item{m}{An adjustR_gof object}

\item{...}{Additional Arguments (not used)}
}
\value{
A list of tibbles
}
\description{
Generates the post-stratification cells based on the provided survey model's
formula and the survey design stratification; e.g., using
\code{svyglm(response ~ gender + race, ...)} with the associated design
\code{svydesign(..., strata = ~region, ...)} will form cells based on the
combination of gender, race, and region. Sample proportions are generated
using the observed number of "topbox" responses in the \eqn{m}th survey mode,
the \eqn{j}th post-stratification cell, and the \eqn{i}th iteration (or year),
defined collectively as \eqn{x_{mij}}, and the sample size for the mode, cell,
and year, defined \eqn{n_{mij}}. The sample proportion of a given cell, mode,
and year is given \eqn{\frac{x_mij}{n_mij}}.
}
\details{
BIMODAL ----------------------------------------------
This function assumes a bimodal analysis. Topbox in this instance is used to
split the data into two separate parts, generating cells and topbox cells
only once. A multimodal analysis will need to loop over a possible list of
topbox values and assign post-stratification cells for each.
}
