% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_gof.R
\name{.validate.gof}
\alias{.validate.gof}
\title{Validate a gof object}
\usage{
.validate.gof(m)
}
\arguments{
\item{m}{a adjustR_gof object to validate}
}
\value{
The validated adjustR_gof object
Throws an error if the model object is invalid
}
\description{
Validate a gof object
}
