% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_spec.R
\name{dplyr.adjustr_spec}
\alias{dplyr.adjustr_spec}
\alias{filter.adjustr_spec}
\alias{arrange.adjustr_spec}
\alias{rename.adjustr_spec}
\alias{select.adjustr_spec}
\alias{slice.adjustr_spec}
\title{\code{dplyr} Methods for \code{adjustr_spec} Objects}
\usage{
\method{filter}{adjustr_spec}(.data, ..., .preserve = FALSE)

\method{arrange}{adjustr_spec}(.data, ...)

\method{rename}{adjustr_spec}(.data, ...)

\method{select}{adjustr_spec}(.data, ...)

\method{slice}{adjustr_spec}(.data, ..., .preserve = FALSE)
}
\arguments{
\item{.data}{the \code{adjustr_spec} object}

\item{...}{additional arguments to underlying method}

\item{.preserve}{as in \code{filter} and \code{slice}}
}
\description{
Core \code{\link[dplyr]{dplyr}} verbs which don't involve grouping
 (\code{\link[dplyr]{filter}}, \code{\link[dplyr]{arrange}},
  \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{select}},
  \code{\link[dplyr]{rename}}, and \code{\link[dplyr]{slice}}) are
implemented and operate on the underlying table of specification parameters.
}
\examples{
\dontrun{
spec = make_spec(eta ~ student_t(df, 0, 1), df=1:10)

arrange(spec, desc(df))
slice(spec, 4:7)
filter(spec, df == 2)
}
}
