% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eightschools_m}
\alias{eightschools_m}
\title{Pre-fitted Eight Schools Model}
\format{
A \code{\link[rstan]{stanfit}} object.
}
\usage{
eightschools_m
}
\description{
A small \code{\link[rstan]{stanfit}} object from the classic
"eight schools" hierarchical model (Rubin, 1981), included for use in
examples and tests.  The model was fit with 2 chains of 1000 iterations
(500 warmup) for a compact object size.
}
\examples{
extract_samp_stmts(eightschools_m)
}
\keyword{datasets}
