% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{extract_model}
\alias{extract_model}
\title{Extract model sampling statements from a Stan model.}
\usage{
extract_model(object)
}
\arguments{
\item{object}{A \code{\link[rstan]{stanfit}} model object.}
}
\value{
Invisbly returns a list of sampling formulas.
}
\description{
Prints a list of sampling statements extracted from the \code{model} block of
a Stan program, with each labelled "parameter" or "data" depending on the
type of variable being sampled.
}
