% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibonacci.R
\name{fib}
\alias{fib}
\title{Calcuate the nth Fibonacci number
This function calculates the nth Fibonacci number using an optimal iterative
method. It has O(n) time complexity}
\usage{
fib(n)
}
\arguments{
\item{n}{An integer representing the position in the Fibonacci sequence.
Must be a non-negative value.}
}
\value{
The nth Fibonacci number
}
\description{
Calcuate the nth Fibonacci number
This function calculates the nth Fibonacci number using an optimal iterative
method. It has O(n) time complexity
}
\examples{
fib(0)
fib(100)
fib(10)
}
