% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantmod.R
\name{gather_symbol}
\alias{gather_symbol}
\title{gather_symbol}
\usage{
gather_symbol(df, symbol = "symbol", index = "index", sep = ".",
  drop = TRUE)
}
\arguments{
\item{df}{A \code{data.frame} or \code{data.table} with a symbol column}

\item{symbol}{The name of the symbol column.}

\item{index}{Columns not to rename. Also, attributes of this column are preserved.}

\item{sep}{Separator between the symbol prefix and the column names.}

\item{drop}{If TRUE (the default), symbols with dates of only NAs will be dropped.}
}
\description{
Create symbol column by peeling prefix from column names

 \code{gather_symbol} creates the symbol column of a 
 \code{data.frame} from the prefix of other columns
 for use after converting from quantmod xts symbols.
}
