% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_gof.R
\name{generate_post_cells.adjustR_gof}
\alias{generate_post_cells.adjustR_gof}
\title{Method to generate post-stratification cells}
\usage{
\method{generate_post_cells}{adjustR_gof}(m, ...)
}
\description{
Method to generate post-stratification cells
}
