% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_gof.R
\name{generate_post_cells}
\alias{generate_post_cells}
\title{Generate post-stratification cells for a adjustR_gof object}
\usage{
generate_post_cells(m, ...)
}
\arguments{
\item{m}{An adjustR_gof object}

\item{...}{Additional Arguments (not used)}
}
\value{
A dataframe of post-stratification cells
}
\description{
BIMODAL ----------------------------------------------
This function assumes a bimodal analysis. Topbox in this instance is used to
split the data into two separate parts, generating cells and topbox cells
only once. A multimodal analysis will need to loop over a possible list of
topbox values and assign post-stratification cells for each.
}
