% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustment_weights.R
\name{get_adjustment_weights}
\alias{get_adjustment_weights}
\title{Compute Pareto-smoothed importance weights for alternative model
specifications}
\usage{
get_adjustment_weights(spec, object, data = NULL)
}
\arguments{
\item{spec}{An object of class \code{adjustr_spec}, probably produced by
\code{\link{make_spec}}, containing the new sampling sampling statements
to replace their counterparts in the original Stan model, and the data,
if any, by which these sampling statements are parametrized.}

\item{object}{A \code{\link[rstan]{stanfit}} model object.}

\item{data}{The data that was used to fit the model in \code{object}.
Required only if one of the new sampling specifications involves Stan data
variables.}
}
\value{
If no \code{specs} were provided, a data frame with columns
  \code{.samp_stmt}, containing the sampling statements, and \code{.weights},
  containing weight objects of class \code{\link[loo]{psis}}. If \code{specs}
  is a data frame or list, then a \code{.weights} entry is added for each row
  or sublist.
}
\description{
Given a set of new sampling statements, which can be parametrized by a
data frame or list, compute Pareto-smoothed importance weights and attach
them to the specification object, for further calculation and plotting.
}
