% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/adjustment.R
\name{getDTSymbols}
\alias{getDTSymbols}
\title{Retrieve stock prices as data.table}
\usage{
getDTSymbols(x, ...)
}
\description{
\code{getDTSymbols} adapts \code{quantmod::getSymbols} to convert
results to \code{data.table}.
}

