% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getfitdistr}
\alias{getfitdistr}
\title{Ajustement avec des distributions continues}
\usage{
getfitdistr(
  x,
  xlab = "Montants de sinistre",
  showplots = TRUE,
  short = TRUE,
  color = list(gamma = "steelblue", lognormal = "firebrick", exponential = "green"),
  plots.as.vars = F
)
}
\arguments{
\item{x}{la variable aléatoire continue a traitè}

\item{xlab}{le label utilisé dans les courbes}

\item{showplots}{valeur booléenne détermine s'il faut générer des courbes}

\item{short}{définit le format de valeur renvoyé}

\item{color}{définit les couleurs des courbes}

\item{plots.as.vars}{Logical détermine s'il faut retourner les courbes}
}
\value{
liste contenant la meilleure distribution ajustée lorsque  \code{short == TRUE}
ou contenant toutes les distributions ajustées autrement avec la meillieure distribution,
si elle existe, dans \code{result$bestdistr} et ceux dont l'hypothèse nulle est acceptée se
trouve dans \code{result$acceptedDistr}
}
\description{
calcule la distribution la mieux ajustée pour les données fournies
}
\examples{
\dontrun{
  (fit <- getfitdistr(x))
}
}
\author{
abdessabour moutik
}
