% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getgoodfit}
\alias{getgoodfit}
\title{Fitting with discrete distributions}
\usage{
getgoodfit(
  x,
  showplots = F,
  short = TRUE,
  plots.as.vars = F,
  xlab = "Number of claims"
)
}
\arguments{
\item{x}{Vector, the discrete random variable.}

\item{showplots}{Logical, determines whether to show plots.}

\item{short}{Logical, determines whether the output will contain the best fitted distribution or all of them.}

\item{plots.as.vars}{Logical, whether to add the plots to the returned list.}

\item{xlab}{a character string to separate the terms.}
}
\value{
A list containing only the best fitted distribution when \code{short=TRUE} otherwise it'll contain all the adjusted distributions.
 If the best fitted distribution exists, it can be accessed through \code{result$bestdistr}, and those who validated the Null
 Hypothesis are found at \code{result$acceptedDistr} ordered from the lowest AIC to the highest.
}
\description{
Calculates the best fitted distribution for the given random variable
}
\examples{
\dontrun{
  (fit <- getgoodfit(x))
}
}
\author{
Abdessabour MOUTIK
}
