% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/adjustment.R
\name{make_raw_return.default}
\alias{make_raw_return.default}
\title{Compute raw return from split adjusted shares, dividends, and closing prices.}
\usage{
\method{make_raw_return}{default}(price_data, split_adjusted_shares,
  split_unadjusted_dividends, initial_rawclose = NULL, ...)
}
\arguments{
\item{price_data}{The historical prices, usually closes, with no adjustments.}

\item{split_adjusted_shares}{Numeric vector of number of shares accounting
for share splits. Typically, starts with 1 and adjusts for each split to reflect what
that 1 share has become.}

\item{split_unadjusted_dividends}{Numeric vector of dividends as they would
have been announced historically. The values returned from Yahoo are split adjusted.
For example, a dividend of 0.2 that occured before a 2:1 split would be reported as
0.1 by Yahoo (assuming no subsequent splits). This function expects the 0.2. Available
as \code{rawdividend} as returned by \code{make_raw_value}.}

\item{initial_rawclose}{Optional starting value to measure returns against. Default is
NULL, i.e., just use the first ticks from the other data vectors. If used, the first
dividend will be included.}

\item{...}{Additional arguments currently unused.}
}
\value{
A numeric vector the same length as the inputs with the arithmetic return from
the first close including dividends.
}
\description{
Computes the return of buying the given number of shares at the first closing price
and exiting at any intermediate closing price including splits and dividends. The
number of shares should only change based on splits; no accounting is done to rebase
the return for other purchases and sales of shares. The dividends must be as they would
historically have been reported, not split adjusted as Yahoo provides. A dividend on
the first tick is not included in the return as Yahoo reports them on the ex-date.
}

