% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustment.R
\name{make_raw_value}
\alias{make_raw_value}
\title{Make additive price adjustments without lookforward bias.}
\usage{
make_raw_value(price_data, splits, dividend, ...)
}
\arguments{
\item{price_data}{Price data for which to compute raw value. Currently in a confused state
where \code{data.table} and \code{xts} are both accepted.}

\item{splits}{Splits as provided by Yahoo, either as \code{data.table}
or \code{xts}.}

\item{dividend}{Split unadjusted dividends as provided by Yahoo, either as \code{data.table}
or \code{xts}.}

\item{...}{Not sure yet.}
}
\value{
A \code{data.table} unless an \code{xts} was passed in containing new
    columns with dividend, splits, rawshares, rawdividend, and rawvalue.
}
\description{
Apply dividend and split data to price series without changing
past data. Changes are applied additively, perhaps multiplicative 
will be added in the future. A dividend paid at time t will increase
the value at time t and all future times. A split at time t will 
apply a factor at time t and all future times. The Rawvalue and
RawShares will be added.
}
\details{
Note that splits and dividends that overlap may not be processed correctly.
Splits are processed first, as the price will be reported post-split, so
we assume the shares have already split, and the dividend is reported 
per share. In practice, this may be false, in which case work would need to be done.

TODO: Need to make sure it works when there is a split or dividend on a date with no price.
}
