% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustment.R
\name{make_reinvested_return.data.table}
\alias{make_reinvested_return.data.table}
\title{Compute reinvested return between start and end.}
\usage{
\method{make_reinvested_return}{data.table}(price_data, start = 1,
  end = nrow(price_data), ...)
}
\arguments{
\item{price_data}{A data.frame of prices as produced by make_raw_data. In particular,
requires columns index, close, rawshares, and rawdividend.}

\item{start}{An index vector (numeric or logical) of starting points.}

\item{end}{An index vector (numeric or logical) of ending poitns.}

\item{...}{Additional arguments.}
}
\value{
A data.frame of reinvested returns with columns index and reinvested_return. 
If the (start,end] ranges do not overlap, the returns are a running calculation from
each start indexed from start+1 to end. If there the (start,end] ranges do not partition
the index of the input, the missing indexes are omitted (especially useful when looking
at trade returns in which you are not always in the market). If there is overlap in
the (start,end] ranges, there is just one return calculated for each end indexed by the 
index at end.
}
\description{
Compute reinvested return between start and end.
}

