% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustment.R
\name{make_shares}
\alias{make_shares}
\title{Turn splits into evolution of single share.}
\usage{
make_shares(splits)
}
\arguments{
\item{splits}{A numeric vector of splits. Most values will be 1. A 2:1 split
is represented as 0.5.}
}
\value{
A numeric vector of shares held if splits are applied to an
initial position of 1 share.
}
\description{
Turn splits into evolution of single share.
}

