% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_spec.R
\name{make_spec}
\alias{make_spec}
\alias{print.adjustr_spec}
\alias{length.adjustr_spec}
\title{Set Up Model Adjustment Specifications}
\usage{
make_spec(...)

\method{print}{adjustr_spec}(x, ...)

\method{length}{adjustr_spec}(x)
}
\arguments{
\item{...}{Ignored.}

\item{x}{An \code{adjustr_spec} object.}
}
\value{
An object of class \code{adjustr_spec}, which is essentially a list
with two elements: \code{samp}, which is a list of sampling formulas, and
\code{params}, which is a list of lists of parameters. Core
\link[=filter.adjustr_spec]{dplyr verbs} which don't involve grouping
(\code{\link[dplyr]{filter}}, \code{\link[dplyr]{arrange}},
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{select}},
\code{\link[dplyr]{rename}}, and \code{\link[dplyr]{slice}}) are
supported and operate on the underlying table of specification parameters.

Invisibly returns \code{x}.

The number of specifications.
}
\description{
Takes a set of new sampling statements, which can be parametrized by other
arguments, data frames, or lists, and creates an \code{adjustr_spec} object
suitable for use in \code{\link{adjust_weights}}.
}
\examples{
make_spec(eta ~ cauchy(0, 1))

make_spec(eta ~ student_t(df, 0, 1), df=1:10)

params = tidyr::crossing(df=1:10, infl=c(1, 1.5, 2))
make_spec(eta ~ student_t(df, 0, 1),
          y ~ normal(theta, infl*sigma),
          params)

}
\seealso{
\code{\link{adjust_weights}}, \code{\link{summarize.adjustr_weighted}}, \code{\link{spec_plot}}
}
