% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primes.R
\name{next_prime}
\alias{next_prime}
\title{Next Prime Number}
\usage{
next_prime(n)
}
\arguments{
\item{n}{A non-negative integer specifying the starting place for the
next prime.}
}
\value{
An integer value for the nearest prime number after \code{n}.
}
\description{
Returns the next prime from a given number \code{n}.
}
\details{
Returns an error if \code{n} is negative.
}
\examples{
next_prime(3)
next_prime(25)
}
