% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_gof.R
\name{run_GOF}
\alias{run_GOF}
\title{Note to self -- implement a join-by argument for the generate post_cells
function. Should pass implicitly.
Wrapper Function}
\usage{
run_GOF(model, topbox_col, topbox_value, mode_col)
}
\arguments{
\item{model}{A survey GLM object}

\item{topbox_col}{Name of the top box column}

\item{topbox_value}{Value indicating the top box in topbox_col}

\item{mode_col}{Name of the mode column}
}
\description{
This is a user friendly wrapper function that translates inputs to the expected
S3 object "adjustR_gof" and calculates Goodness-of-Fit over that object.
}
