% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_gof.R
\name{run_test}
\alias{run_test}
\title{Implement Goodness-of-Fit test}
\usage{
run_test(m, ...)
}
\arguments{
\item{m}{An adjustR_gof object}
}
\description{
The main function associated with the Goodness-of-Fit hypothesis test.
Implements the adjustR_gof methods to process data, calculate observed
test statistics and simulations. Modifies the adjustR_gof object in place
and returns the object invisbly.
}
