% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primes.R
\name{sieve}
\alias{sieve}
\title{Sieve of Eratosthenes}
\usage{
sieve(n)
}
\arguments{
\item{n}{An integer specifying the upper limit (exclusive) for finding prime
numbers.}
}
\value{
A vector of prime numbers less than \code{n}.
}
\description{
Return a vector of prime numbers less than the given number using the
sieve of Eratosthenes method
}
\details{
Throws an error if \code{n} is less than 2.
}
\examples{
sieve(10)
sieve(20)
}
