% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_weights.R
\name{spec_plot}
\alias{spec_plot}
\title{Plot Posterior Quantities of Interest Under Alternative Model Specifications}
\usage{
spec_plot(
  x,
  by,
  post,
  only_mean = FALSE,
  ci_level = 0.8,
  outer_level = 0.95,
  ...
)
}
\arguments{
\item{x}{An \code{adjustr_weighted} object.}

\item{by}{The x-axis variable, which is usually one of the specification
parameters. Can be set to \code{1} if there is only one specification.
Automatically quoted and evaluated in the context of \code{x}.}

\item{post}{The posterior quantity of interest, to be computed for each
resampled draw of each specificaiton. Should evaluate to a single number
for each draw. Automatically quoted and evaluated in the context of \code{x}.}

\item{only_mean}{Whether to only plot the posterior mean. May be more stable.}

\item{ci_level}{The inner credible interval to plot. Central
100*ci_level% intervals are computed from the quantiles of the resampled
posterior draws.}

\item{outer_level}{The outer credible interval to plot.}

\item{...}{Ignored.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object which can be further
  customized with the \strong{ggplot2} package.
}
\description{
Uses weights computed in \code{\link{adjust_weights}} to plot posterior
quantities of interest versus specification parameters
}
\examples{
\dontrun{
spec = make_spec(eta ~ student_t(df, 0, scale),
                 df=1:10, scale=seq(2, 1, -1/9))
adjusted = adjust_weights(spec, eightschools_m)

spec_plot(adjusted, df, theta[1])
spec_plot(adjusted, df, mu, only_mean=TRUE)
spec_plot(adjusted, scale, tau)
}

}
