% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantmod.R
\name{spread_symbol}
\alias{spread_symbol}
\title{spread_symbol}
\usage{
spread_symbol(df, symbol = "symbol", index = "index", sep = ".",
  drop = TRUE)
}
\arguments{
\item{df}{A \code{data.frame} or \code{data.table} with a symbol column}

\item{symbol}{The name of the symbol column.}

\item{index}{Columns not to rename. Also, attributes of this column are preserved.}

\item{sep}{Separator between the symbols and the column names.}

\item{drop}{If TRUE (the default), symbols with price columns of only NAs will be dropped.}
}
\description{
Spread symbol column across numeric columns

 \code{spread_symbol} moves the symbol column of a 
 \code{data.frame} to prefix the numeric columns
 so it is not dropped when converting to xts.
}
