% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_weights.R
\name{summarise.adjustr_weighted}
\alias{summarise.adjustr_weighted}
\alias{summarize.adjustr_weighted}
\title{Summarize Posterior Distributions Under Alternative Model Specifications}
\usage{
\method{summarise}{adjustr_weighted}(.data, ..., .resampling = F, .model_data = NULL)

\method{summarize}{adjustr_weighted}(.data, ..., .resampling = F, .model_data = NULL)
}
\arguments{
\item{.data}{An \code{adjustr_weighted} object.}

\item{...}{Name-value pairs of expressions. The name of each argument will be
the name of a new variable, and the value will be computed for the
posterior distribution of eight alternative specification. For example,
a value of \code{mean(theta)} will compute the posterior mean of
\code{theta} for each alternative specification.}

\item{.resampling}{Wether to compute summary statistics by first resampling
the data according to the weights. Defaults to \code{FALSE}, but will be
used for any summary statistic that is not \code{mean}, \code{var} or
\code{sd}.}

\item{.model_data}{Stan model data, if not provided in the earlier call to
\code{\link{adjust_weights}}.}
}
\value{
An \code{adjustr_weighted} object, wth the new columns specified in
\code{...} added.
}
\description{
Uses weights computed in \code{\link{adjust_weights}} to compute posterior
summary statistics. These statistics can be compared against their reference
values to quantify the sensitivity of the model to aspects of its
specification.
}
