% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/adjustData.R
\name{true_value}
\alias{true_value}
\title{Make additive price adjustments without lookforward bias.}
\usage{
true_value(price_data, dividend, ..., splits)
}
\description{
Apply dividend and split data to price series without changing
past data. Changes are applied additively, perhaps multiplicative
will be added in the future. A dividend paid at time t will increase
the value at time t and all future times. A split at time t will
apply a factor at time t and all future times. The TrueValue and
TrueShares will be added.
}
\details{
Because one could accidentally pass in the dividend object as the split and
vice versa, split must be assigned by name.

Note that splits and dividends that overlap may not be processed correctly.
Splits are processed first, as the price will be reported post-split, so
we assume the shares have already split, and the dividend is reported
per share. In practice, this may be false, in which case work would need to be done.

TODO: Need to make sure it works when there is a split or dividend on a date with no price.
}

