% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustment.R
\name{unadjust.xts}
\alias{unadjust.xts}
\title{Turn split adjusted dividends into unadjusted dividends.}
\usage{
\method{unadjust}{xts}(split_adjusted_dividend, splits, ...)
}
\arguments{
\item{split_adjusted_dividend}{An \code{xts} of dividends, either 1 column or with a column name
matching dividend.}

\item{splits}{An \code{xts} of splits, either 1 column or with a column \code{spl}.}

\item{...}{Additional arguments passed to unadjust.default, for example, \code{max_decimals}.}
}
\value{
An \code{xts} of merged splits and dividends with new fields \code{shares}
and \code{unadjusted_dividend}. Missing splits are filled in with 1. Missing dividends are filled in with 0.
}
\description{
Assumes that unadjusted dividends are only accurate to 3 decimal places
unless \code{max_decimals} specifies otherwise.
}

