% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_gof.R
\name{validate_adjustR_gof}
\alias{validate_adjustR_gof}
\title{Validate a gof object}
\usage{
validate_adjustR_gof(m)
}
\arguments{
\item{m}{a adjustR_gof object to validate}
}
\value{
The validated adjustR_gof object
Throws an error if the model object is invalid
}
\description{
Validate a gof object
}
