% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xts.R
\name{as.xts.data.table}
\alias{as.xts.data.table}
\title{as.xts.data.table}
\usage{
as.xts.data.table(x, ...)
}
\arguments{
\item{x}{data.table to convert to xts, must have \emph{POSIXct} in field called "index".}

\item{...}{additional arguments to pass to xts constructor such as attributes, but not
order.by.}
}
\description{
convert data.table to xts. symbol column
will be joined to nonindex field names.

\code{as.xts} is specialized in \code{adjustr}
to transform an \code{xts} indexed by a date to 
a \code{data.table} with a Date field.
}

