% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_weights.R
\name{extract_samp_stmts}
\alias{extract_samp_stmts}
\title{Extract Model Sampling Statements From a Stan Model.}
\usage{
extract_samp_stmts(object)
}
\arguments{
\item{object}{A \code{\link[rstan]{stanfit}} model object.}
}
\value{
Invisbly returns a list of sampling formulas.
}
\description{
Prints a list of sampling statements extracted from the \code{model} block of
a Stan program, with each labelled "parameter" or "data" depending on the
type of variable being sampled.
}
\examples{
\dontrun{
extract_samp_stmts(eightschools_m)
#> Sampling statements for model 2c8d1d8a30137533422c438f23b83428:
#>   parameter   eta ~ std_normal()
#>   data        y ~ normal(theta, sigma)
}
}
