% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustment.R
\name{getDTSymbols}
\alias{getDTSymbols}
\title{Retrieve stock prices as data.table}
\usage{
getDTSymbols(x, ..., cache = TRUE)
}
\arguments{
\item{x}{A symbol or list of symbols to retrieve from Yahoo and adjust.}

\item{...}{Additional arguments passed to quantmod::getSymbols.}

\item{cache}{Whether to check cache for data before calling quantmod::getSymbols.}
}
\value{
A \code{data.table} of data for input symbols with adjustments.
}
\description{
\code{getDTSymbols} adapts \code{quantmod::getSymbols} to convert
results to \code{data.table}.
}
