% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getfitdistr}
\alias{getfitdistr}
\title{Fitting with continuous distributions}
\usage{
getfitdistr(
  x,
  xlab = "Claim amounts",
  showplots = TRUE,
  short = TRUE,
  color = list(gamma = "steelblue", lognormal = "firebrick", exponential = "green"),
  plots.as.vars = F
)
}
\arguments{
\item{x}{Vector, the continuous random variable.}

\item{xlab}{a character string to separate the terms.}

\item{showplots}{Logical, determines whether to show plots.}

\item{short}{Logical, determines whether the output will contain the best fitted distribution or all of them.}

\item{color}{Named Vector, the colors to use in the plot.}

\item{plots.as.vars}{Logical, whether to add the plots to the returned list.}
}
\value{
A list containing only the best fitted distribution when \code{short=TRUE} otherwise it'll contain all the adjusted distributions.
If the best fitted distribution exists, it can be accessed through \code{result$bestdistr}, and those who validated the Null
Hypothesis are found at \code{result$acceptedDistr} ordered from the lowest AIC to the highest.
}
\description{
Calculates the best fitted distribution for the given random variable
}
\examples{
\dontrun{
  (fit <- getfitdistr(x))
}
}
\author{
Abdessabour MOUTIK
}
