% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getgoodfit}
\alias{getgoodfit}
\title{Ajustement avec des distributions discrètes}
\usage{
getgoodfit(
  x,
  showplots = F,
  short = TRUE,
  plots.as.vars = F,
  xlab = "Nombre de sinistres"
)
}
\arguments{
\item{x}{la variable aléatoire discrète a traitè}

\item{showplots}{valeur booléenne détermine s'il faut afficher les courbes}

\item{short}{définit le format de la valeur renvoyée par défaut TRUE}

\item{plots.as.vars}{Logical détermine s'il faut retourner les courbes}

\item{xlab}{le label utilisé dans les courbes}
}
\value{
liste contenant la meilleure distribution ajustée lorsque  \code{short == TRUE}
ou contenant toutes les distributions ajustées autrement avec la meillieure distribution,
si elle existe, dans \code{result$bestdistr} et ceux dont l'hypothèse nulle est acceptée se
trouve dans \code{result$acceptedDistr}
}
\description{
calcule la distribution la mieux ajustée pour les données fournies
}
\examples{
\dontrun{
   (fit <- getgoodfit(x))
}
}
\author{
abdessabour moutik
}
