% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/adjustment.R
\name{make_raw_return.data.table}
\alias{make_raw_return.data.table}
\title{Compute raw return between start and end.}
\usage{
\method{make_raw_return}{data.table}(price_data, start = 1,
  end = nrow(price_data), ...)
}
\arguments{
\item{price_data}{A \code{data.table} of price data as produced by \code{make_raw_data}.
In particular, it must have columns \code{close, rawshares, rawdividend}.}

\item{start}{An index vector (numeric or logical) of starting points.}

\item{end}{An index vector (numeric or logical) of ending poitns.}

\item{...}{Additional arguments.}
}
\value{
A \code{data.table} of raw returns with columns index and rawreturn.
If the (start,end] ranges do not overlap, the returns are a running calculation from
each start indexed from start+1 to end. If there the (start,end] ranges do not partition
the index of the input, the missing indexes are omitted (especially useful when looking
at trade returns in which you are not always in the market). If there is overlap in
the (start,end] ranges, there is just one return calculated for each end indexed by the
index at end.
}
\description{
Compute raw return between start and end.
}

