% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/adjustment.R
\name{make_raw_return}
\alias{make_raw_return}
\title{Compute period returns as value change plus dividends.}
\usage{
make_raw_return(x, ...)
}
\description{
Compute period returns as value change plus dividends.
}

