% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/adjustment.R
\name{make_raw_return}
\alias{make_raw_return}
\title{Compute period returns as value change plus dividends.}
\usage{
make_raw_return(price_data, ...)
}
\arguments{
\item{price_data}{Price data for which to compute raw returns.}

\item{...}{Additional arguments.}
}
\value{
Return data assuming shares are split according to split data
    and dividends are paid, one for each share, but not reinvested.
}
\description{
Compute period returns as value change plus dividends.
}

