% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_spec.R
\name{make_spec}
\alias{make_spec}
\title{Create a model specification object}
\usage{
make_spec(...)
}
\arguments{
\item{...}{Model specification. Each argument can either be a formula,
  a named vector, data frames,
  or lists.


data frame, or list of lists, containing parameters which will
  be substituted into the sampling statements contained in \code{spec_stamp}.
  If a data frame, columns will be substituted. If a list of lists, the named
  entries of each sublist will be substituted.
spec_samp A formula, or list of formulas, of new sampling statements
  to replace their counterparts in the original Stan model. Sampling
  distributions can be parametrized by constants, Stan parameters and data,
  or by variables from the \code{specs} object.}
}
\value{
An object of class \code{adjustr_spec}, which is essentially a list
with two elements: \code{samp}, which is a list of sampling formulas, and
\code{params}, which is a list of lists of parameters. Core
\link[=filter.adjustr_spec]{dplyr verbs} which don't involve grouping
(\code{\link[dplyr]{filter}}, \code{\link[dplyr]{arrange}},
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{select}},
\code{\link[dplyr]{rename}}, and \code{\link[dplyr]{slice}}) are
supported and operate on the underlying table of specification parameters.
}
\description{
Takes a set of new sampling statements, which can be parametrized by other
arguments, data frames, or lists, and creates an \code{adjustr_spec} object
suitable for use in \code{\link{adjustment_weights}}.
}
