% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_weights.R
\name{pull.adjustr_weighted}
\alias{pull.adjustr_weighted}
\title{Extract Weights From an \code{adjustr_weighted} Object}
\usage{
\method{pull}{adjustr_weighted}(.data, var = ".weights")
}
\arguments{
\item{.data}{A table of data}

\item{var}{A variable, as in \code{\link[dplyr]{pull}}. The default returns
the \code{.weights} column, and if there is only one row, it returns the
first element of that column}
}
\description{
This function modifies the default behavior of \code{dplyr::pull} to extract
the \code{.weights} column.
}
