% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/adjustment.R
\name{unadjust.data.table}
\alias{unadjust.data.table}
\title{Turn split adjusted dividends into unadjusted dividends.}
\usage{
\method{unadjust}{data.table}(split_adjusted_dividend, splits, ...)
}
\arguments{
\item{split_adjusted_dividend}{A data.table of with a \code{dividend} column and
an \code{index} column to merge on.}

\item{splits}{A \code{data.table} of with columns \code{splits} and \code{index}.}

\item{...}{Additional arguments passed to unadjust.default, for example, \code{max_decimals}.}
}
\value{
A data.table of merged splits and dividends with new fields \code{shares}
and {unadjusted_dividend}. The merge matches on the \code{index} field. Missing
}
\description{
Assumes that unadjusted dividends are only accurate to 3 decimal places
unless \code{max_decimals} specifies otherwise.
}

