% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustment.R
\name{unadjust}
\alias{unadjust}
\title{Turn split adjusted dividends into unadjusted dividends.}
\usage{
unadjust(split_adjusted_dividend, splits, ...)
}
\arguments{
\item{split_adjusted_dividend}{Split adjusted dividends for a stock.}

\item{splits}{Stock split data.}

\item{...}{Additional arguments.}
}
\description{
Turn split adjusted dividends into unadjusted dividends.
}
